+++ CHANGELOG.md
@@ -15,7 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Breaking changes
 
-None.
+- `extract::extractor_middleware::ExtractorMiddlewareResponseFuture` moved
+  to `extract::extractor_middleware::future::ResponseFuture`
+- `routing::BoxRouteFuture` moved to `routing::future::BoxRouteFuture`
+- `routing::EmptyRouterFuture` moved to `routing::future::EmptyRouterFuture`
+- `routing::RouteFuture` moved to `routing::future::RouteFuture`
 
 # 0.1.2 (01. August, 2021)